requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2025.3
